+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0.3] - Unreleased
+## [0.0.3] - 2025-09-21
 
 ### Added
 - Filtering
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Sorting: option method `withFileSort` renamed in `sort`
-- Child limit: renaming from 'children limit'
+- Child limit: various renaming and refactor
 
 
 ## [0.0.2] - 2025-09-16